added shop, map moved
